config.py
"""
Configuration and environment variable loading.
"""
import os
from dotenv import load_dotenv

load_dotenv()

IMAP_HOST = os.getenv("IMAP_HOST")
IMAP_USER = os.getenv("IMAP_USER")
IMAP_PASSWORD = os.getenv("IMAP_PASSWORD")
IMAP_FOLDER = 'INBOX'

CHAT_MODEL = 'qwen3:8b'
